// ts/commander/src/api/schemas/sectionWithSubsections.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Corpora API
 * API for managing and processing corpora
 * OpenAPI spec version: 0.1.0
 */
import type { SectionWithSubsectionsIntroduction } from "./sectionWithSubsectionsIntroduction";
import type { SectionWithSubsectionsInstructions } from "./sectionWithSubsectionsInstructions";
import type { SectionWithSubsectionsOrder } from "./sectionWithSubsectionsOrder";
import type { SubsectionOut } from "./subsectionOut";

export interface SectionWithSubsections {
  title: string;
  introduction?: SectionWithSubsectionsIntroduction;
  instructions?: SectionWithSubsectionsInstructions;
  order?: SectionWithSubsectionsOrder;
  id: string;
  project_id: string;
  created_at: string;
  updated_at: string;
  subsections: SubsectionOut[];
}