// ts/commander/src/api/schemas/subsectionOut.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Corpora API
 * API for managing and processing corpora
 * OpenAPI spec version: 0.1.0
 */
import type { SubsectionOutContent } from "./subsectionOutContent";
import type { SubsectionOutInstructions } from "./subsectionOutInstructions";
import type { SubsectionOutOrder } from "./subsectionOutOrder";

export interface SubsectionOut {
  title: string;
  content?: SubsectionOutContent;
  instructions?: SubsectionOutInstructions;
  order?: SubsectionOutOrder;
  id: string;
  section_id: string;
  created_at: string;
  updated_at: string;
}